refactored the login page
